requirements.txt
-i https://pypi.org/simple
aioredis==2.0.1
alembic==1.11.3
amqp==5.1.1 ; python_version >= '3.6'
annotated-types==0.5.0 ; python_version >= '3.7'
anyio==3.7.1 ; python_version >= '3.7'
async-timeout==4.0.3 ; python_version >= '3.7'
asyncpg==0.28.0
autoflake==2.2.0
bcrypt==4.0.1
billiard==4.1.0 ; python_version >= '3.7'
black==23.7.0
celery==5.3.1
certifi==2023.7.22 ; python_version >= '3.6'
click==8.1.7 ; python_version >= '3.7'
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0 ; python_version >= '3.6'
colorama==0.4.6 ; sys_platform == 'win32'
dnspython==2.4.2 ; python_version >= '3.8' and python_version < '4.0'
ecdsa==0.18.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==2.0.0.post2
exceptiongroup==1.1.3 ; python_version < '3.11'
fastapi==0.103.0
fastapi-cache2[redis]==0.2.1
fastapi-versioning==0.10.0
flake8==6.1.0
flower==2.0.1
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.17.3 ; python_version >= '3.7'
httpx==0.24.1
humanize==4.8.0 ; python_version >= '3.8'
idna==3.4 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
isort==5.12.0
itsdangerous==2.1.2
jinja2==3.1.2 ; python_version >= '3.7'
jose==1.0.0
kombu==5.3.1 ; python_version >= '3.8'
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
mccabe==0.7.0 ; python_version >= '3.6'
mypy-extensions==1.0.0 ; python_version >= '3.5'
nodeenv==1.8.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
packaging==23.1 ; python_version >= '3.7'
passlib==1.7.4
pathspec==0.11.2 ; python_version >= '3.7'
pendulum==2.1.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pillow==10.0.0
platformdirs==3.10.0 ; python_version >= '3.7'
pluggy==1.3.0 ; python_version >= '3.8'
prometheus-client==0.17.1 ; python_version >= '3.6'
prompt-toolkit==3.0.39 ; python_full_version >= '3.7.0'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycodestyle==2.11.0 ; python_version >= '3.8'
pydantic[email]==2.3.0
pydantic-core==2.6.3 ; python_version >= '3.7'
pydantic-settings==2.0.3
pyflakes==3.1.0 ; python_version >= '3.8'
pyright==1.1.324
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0 ; python_version >= '3.8'
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6 ; python_version >= '3.7'
pytz==2023.3
pytzdata==2020.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
redis==4.6.0
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
setuptools==68.1.2 ; python_version >= '3.8'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqladmin==0.14.1
sqlalchemy==2.0.20
starlette==0.27.0 ; python_version >= '3.7'
tomli==2.0.1 ; python_version < '3.11'
tornado==6.3.3 ; python_version >= '3.8'
typing-extensions==4.7.1 ; python_version >= '3.7'
tzdata==2023.3 ; python_version >= '2'
uvicorn==0.23.2
vine==5.0.0 ; python_version >= '3.6'
wcwidth==0.2.6
wtforms==3.0.1 ; python_version >= '3.7'
